+++ src/server/middlewares/fetch.ts
@@ -4,14 +4,6 @@ import { getFileData, getUrlData } from '../utils/fetch';
 export default (req, res, next) => {
   const locals = res.locals as Locals;
   const routeConfig = locals.routeConfig;
-
-  if (!routeConfig.fetch || !routeConfig._request?.url) return next();
-
-  if (routeConfig.fetchCount == 0) {
-    locals.data = routeConfig.fetchData ?? (routeConfig.skipFetchError ? routeConfig.mock : routeConfig.fetchError);
-    return next();
-  }
-
   if (routeConfig._request?.url?.startsWith("http")) {
     _FetchUrl(req, res, next)
   } else {
@@ -22,23 +14,23 @@ const _FetchUrl = async (_req, res, next) => {
   const locals = res.locals as Locals;
   const routeConfig = locals.routeConfig;
 
-  const { _request } = routeConfig
+  const _request = routeConfig._request;
   if (!_request?.url?.startsWith("http")) return next();
 
   if (routeConfig.fetchCount == 0) {
-    locals.data = routeConfig.fetchData ?? (routeConfig.skipFetchError ? routeConfig.mock : routeConfig.fetchError);
+    const { isError, response } = routeConfig.fetchData || {};
+    locals.data = isError ? (routeConfig.skipFetchError ? routeConfig.mock : response) : response;
     return next();
   }
 
-  const { fetchData, fetchError } = await getUrlData(_request!);
+  const fetchData = await getUrlData(_request!);
   routeConfig.fetchData = fetchData;
-  routeConfig.fetchError = fetchError;
   routeConfig.fetchCount!--;
   delete routeConfig.store?.["_IterateResponse"];
   delete routeConfig.store?.["_IterateRoutes"];
   delete routeConfig.store?.["_CrudOperation"];
 
-  locals.data = fetchData ?? (routeConfig.skipFetchError ? routeConfig.mock : fetchError);
+  locals.data = fetchData.isError ? (routeConfig.skipFetchError ? routeConfig.mock : fetchData.response) : fetchData.response;
 
   next();
 }
@@ -47,22 +39,22 @@ const _FetchFile = (_req, res, next) => {
   const routeConfig = locals.routeConfig;
 
   const { _request, _extension = '' } = routeConfig;
-  if (_request?.url?.startsWith("http") || ![".json", ".har", ".txt"].includes(_extension)) return next();
+  if (_request?.url?.startsWith("http") || ![".json", ".jsonc", ".har", ".txt"].includes(_extension)) return next();
 
   if (routeConfig.fetchCount == 0) {
-    locals.data = routeConfig.fetchData ?? (routeConfig.skipFetchError ? routeConfig.mock : routeConfig.fetchError);
+    const { isError, response } = routeConfig.fetchData || {};
+    locals.data = isError ? (routeConfig.skipFetchError ? routeConfig.mock : response) : response;
     return next();
   }
 
-  const { fetchData, fetchError } = getFileData(_request!.url!, _extension);
+  const fetchData = getFileData(_request!.url!, _extension);
   routeConfig.fetchData = fetchData;
-  routeConfig.fetchError = fetchError;
   routeConfig.fetchCount!--;
   delete routeConfig.store?.["_IterateResponse"];
   delete routeConfig.store?.["_IterateRoutes"];
   delete routeConfig.store?.["_CrudOperation"];
 
-  locals.data = fetchData ?? (routeConfig.skipFetchError ? routeConfig.mock : fetchError);
+  locals.data = fetchData.isError ? (routeConfig.skipFetchError ? routeConfig.mock : fetchData.response) : fetchData.response;
 
   next();
 }